refactor(pointer): split motion handling into focused submodules
Reorganize input/pointer/motion into a folder-based module layout to
improve cohesion and make the motion path easier to maintain.

- keep mod.rs as the entry point for handle_pointer_motion_absolute
- add routing.rs for monitor resolution and motion dispatch
- add bloom.rs for bloom drag and pull-to-detach behavior
- add drag.rs for node drag lifecycle handling
- add focus.rs for hover-focus behavior
- add cluster.rs for cluster join candidate and dwell logic
- add pending.rs for press-to-drag transition helpers
- add tests.rs for motion-specific unit tests

Also cleaned up imports and dead code surfaced during the split.

Verified with:
- cargo check
